Researcher-AI-Agent/requirements.txt
ollama
ddgs
python-dotenv